Refactor readFromFile function to use fs.open
This commit refactors the readFromFile function in our file handling module. Previously, we were using fs.readFile to read from a file. This method opens, reads, and then closes the file. However, in some cases, we might want to perform multiple operations on a file, and constantly opening and closing it can be inefficient.

To address this, we've changed the readFromFile function to use fs.open instead. This method opens the file and returns a file handle, which can be used for subsequent operations. We then use the file handle's readFile method to read the file. Once we're done with the file, we make sure to close it using the file handle's close method. This is done in a finally block to ensure that the file is closed even if an error occurs.

This change will make our file operations more efficient and give us more control over when the file is opened and closed.